+++ src/antity-pgsql.ts
@@ -1,4 +1,4 @@
 export { SQLEntity } from './entity';
 export { filter } from './filter/filter';
 export { execute } from './crud/execute';
-export type { Row, PGClient, PGResponse, SelectResponse, Filters, Filter, MatchMode, Sort, Operation, Geometry, MappedType } from './types';
+export type { Row, PGClient, PGResponse, SelectResponse, Filters, Filter, SqlValue, MatchMode, Sort, Operation, Geometry, MappedType } from './types';