+++ src/crud/select.ts
@@ -2,7 +2,7 @@ import { deleteProps } from "@dwtechs/sparray";
 import { execute as exe } from "./execute";
 import { quoteIfUppercase } from "./quote";
 import { filter } from "../filter/filter";
-import type { PGResponse, SelectResponse, Filter, Filters, Sort, PGClient, LogicalOperator } from "../types";
+import type { PGResponse, SelectResponse, SqlValue, Filters, Sort, PGClient, LogicalOperator } from "../types";
 
 export class Select {
   
@@ -28,7 +28,7 @@ export class Select {
     sortOrder: Sort | null = null,
     filters: Filters | null = null,
     operator: LogicalOperator = "AND"
-  ): { query: string, args: (Filter["value"])[] } {
+  ): { query: string, args: SqlValue[] } {
     const p = limit ? this._count : '';
     const c = this._cols ? this._cols : '*';
     const baseQuery = `SELECT ${c}${p} FROM ${quoteIfUppercase(schema)}.${quoteIfUppercase(table)}`;
@@ -43,7 +43,7 @@ export class Select {
 
   public execute(
     query: string,
-    args: (Filter["value"])[],
+    args: SqlValue[],
     client: PGClient | null,
   ): Promise<SelectResponse> {